Reorganized numcosmo app.
